+++ atest/examples/example.jl
@@ -1,0 +1,4 @@
+add_together(x,y) = x + y
+
+
+add_together(1)