+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - features
 
+  - added experimental detection of Julia and Jedi language servers ([#481])
   - make the extension work with `jupyterlab-classic` - experimental, not all features are functional yet ([#465])
   - new status "Server extension missing" and a dialog with advice was added to help users with atypical configurations ([#476])
   - for developers: the verbosity of console logs is now controllable from settings and set to warn by default ([#480])
@@ -24,6 +25,7 @@
 [#478]: https://github.com/krassowski/jupyterlab-lsp/pull/478
 [#479]: https://github.com/krassowski/jupyterlab-lsp/pull/479
 [#480]: https://github.com/krassowski/jupyterlab-lsp/pull/480
+[#481]: https://github.com/krassowski/jupyterlab-lsp/pull/481
 
 ### `jupyter-lsp 1.0.1` (unreleased)